+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [v0.3.0] - 2020-04-18
+
+### Added
+- `app` directory: Demo project to test `react-sac` component
+- `react-sac` directory: The component's folder
+- `./react-sac/index.d.ts`: Typescript typings
+
 ## [v0.2.0] - 2020-04-17
 
 ### Added
@@ -19,4 +26,5 @@
 	- Provide selection callbacks
 
 [Unreleased]: https://github.com/kapantzak/react-sac/compare/master...develop
+[v0.3.0]: https://github.com/kapantzak/react-sac/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/kapantzak/react-sac/compare/v0.1.0...v0.2.0